+++ docs/RELEASE-NOTES.rst
@@ -51,6 +51,7 @@ Bug Fixes
 * `Issue 2632 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/2632>`_: Fix hubmode support with NodePortLocal
 * `Issue 2821 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/2821>`_: Fix for additionalVirtualAddresses with serviceAddress config
 * `Issue 2550 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/2550>`_: Ability to specify monitors for TransportServer CR
+* `Issue 2912 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/2912>`_: Enable metrics with ipv6 mode
 * Fix for recreating the LTM objects when CIS restarts in IPAM mode.
 * Improved error handling for GTM objects with cccl-gtm-agent.
 * Fix crash issue with liveness probe in NextGen routes